validation remote to separate file
