+++ helpers/index.js
@@ -1,5 +1,5 @@
 const HttpError = require("./HttpError");
 
 module.exports = {
-    HttpError,
+    HttpError
 }
