+++ models/contacts.json
@@ -73,7 +73,7 @@
   },
   {
     "id": "QAugLPeTkoiU50-1GFKiN",
-    "name": "AFtor",
+    "name": "Genry Greyzer",
     "email": "elementum@scelerisquescelerisquedui.net",
     "phone": "(748) 206-2688"
   },
